+++ Source/Data/Library/Entities/EncounterWall.cs
@@ -3,6 +3,6 @@ namespace VttTools.Data.Library.Entities;
 public class EncounterWall {
     public Guid EncounterId { get; set; }
     public Encounter Encounter { get; set; } = null!;
-    public uint Index { get; set; }
+    public ushort Index { get; set; }
     public List<EncounterWallSegment> Segments { get; set; } = [];
 }
