+++ Source/Domain/Library/Encounters/ApiContracts/EncounterAssetBulkCloneRequest.cs
@@ -1,5 +1,6 @@
 namespace VttTools.Library.Encounters.ApiContracts;
 
-public record EncounterAssetBulkCloneRequest : Request {
-    public required List<uint> Indices { get; init; }
+public record EncounterAssetBulkCloneRequest
+    : Request {
+    public required List<ushort> Indices { get; init; }
 }
