+++ Source/Domain/Library/Encounters/ApiContracts/EncounterAssetBulkDeleteRequest.cs
@@ -1,5 +1,6 @@
 namespace VttTools.Library.Encounters.ApiContracts;
 
-public record EncounterAssetBulkDeleteRequest : Request {
-    public required List<uint> Indices { get; init; }
+public record EncounterAssetBulkDeleteRequest
+    : Request {
+    public required List<ushort> Indices { get; init; }
 }
