+++ Source/Domain/Library/Encounters/Model/EncounterAsset.cs
@@ -2,7 +2,7 @@
 
 public record EncounterAsset {
     public Guid AssetId { get; init; }
-    public uint Index { get; init; }
+    public ushort Index { get; init; }
 
     [MaxLength(128)]
     public string? Name { get; init; }
