+++ Source/Domain/Library/Encounters/Model/EncounterWall.cs
@@ -1,6 +1,6 @@
 namespace VttTools.Library.Encounters.Model;
 
 public record EncounterWall {
-    public uint Index { get; init; }
+    public ushort Index { get; init; }
     public IReadOnlyList<EncounterWallSegment> Segments { get; init; } = [];
 }
