+++ Source/Domain/Library/Encounters/Model/EncounterWallSegment.cs
@@ -1,7 +1,7 @@
 namespace VttTools.Library.Encounters.Model;
 
 public record EncounterWallSegment {
-    public uint Index { get; init; }
+    public ushort Index { get; init; }
     [MaxLength(128)]
     public string? Name { get; init; }
     public Pole StartPole { get; init; } = new(0, 0, 0);
